02_rag/pyproject.toml
[project]
name = "02-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.13",
    "llama-index-core>=0.12.44",
    "llama-index-embeddings-ollama>=0.6.0",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-deepseek>=0.1.2",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-llms-openai>=0.4.7",
    "llama-index-vector-stores-chroma>=0.4.2",
    "pymupdf>=1.26.1",
]
